Extract PresetUIManager from MainComponent
- Create separate PresetUIManager component for all preset-related functionality
- Move bank/preset ComboBoxes, save button, and all preset dialogs to PresetUIManager
- Update MainComponent to use PresetUIManager as a contained component
- Delegate preset-related ValueTree property changes to PresetUIManager
- Add PresetUIManager to CMakeLists.txt build configurations
- Maintain exact same preset functionality with improved code organization

Part of Phase 2 MainComponent refactoring strategy.
