+++ src/CMakeLists.txt
@@ -51,6 +51,7 @@ target_sources(YMulator-Synth
         ui/RotaryKnob.cpp
         ui/EnvelopeDisplay.cpp
         ui/AlgorithmDisplay.cpp
+        ui/PresetUIManager.cpp
         utils/PresetManager.cpp
         utils/VOPMParser.cpp
         # ymfm library sources
